# Faster diffusion