--- a/emergence_lib/Cargo.toml
+++ b/emergence_lib/Cargo.toml
@@ -30,6 +30,8 @@ itertools = "0.10.5"
 bevy_screen_diagnostics = "0.2"
 anyhow = "1.0.69"
 serde_json = "1.0.94"
+hashbrown = { version = "0.12", features = ["rayon"] }
+rayon = "1.7.0"
 
 [dev-dependencies]
 criterion = "0.4"