Fixes #718, at least so far.

> signal_diffusion_modest time:   [26.909 ms 27.052 ms 27.205 ms]
                        change: [-64.437% -64.238% -64.017%] (p = 0.00 < 0.05)
                        Performance has improved.

For the rayon step.

> signal_diffusion_modest time:   [26.345 ms 26.508 ms 26.684 ms]
                        change: [-2.8006% -2.0121% -1.1968%] (p = 0.00 < 0.05)
                        Performance has improved.

For inlining.

We still need to improve performance dramatically, but it appears that shadows are the big offender currently.